+++ app/javascript/utilities/utilities.ts
@@ -248,4 +248,4 @@ export const baseMorale = 15
 export const baseRally = 12
 export const titleName = "A Hex Too Far"
 export const subtitleName = "Light Tactical Battle System"
-export const serverVersion = "0.28"
+export const serverVersion = "0.29"