+++ app/models/utility/server.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Utility
+  class Server
+    VERSION = "0.29"
+  end
+end